requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
python-dotenv
pytest